requirements.txt
# loading environment variables
python-dotenv==0.20.0

# managing tweets
tweepy==4.10.0

# making api calls
requests==2.28.1

# paraphrasing quotes
transformers==4.20.1
tensorflow==2.9.1
sentencepiece==0.1.96
sacremoses==0.0.53
torch==1.12.0
protobuf==3.19.4
